+++ src/main/java/com/filezipper/iostreams/FileInputStream.java
@@ -12,8 +12,8 @@ public final class FileInputStream implements  IInputStream{
     }
 
     @Override
-    public Byte read() throws IOException{
-        return (byte) fis.read();
+    public int read() throws IOException{
+        return fis.read();
     }
     @Override
     public Byte[] readNBytes(Integer noOfBytes) throws IOException{
