+++ src/main/java/com/filezipper/iostreams/IInputStream.java
@@ -3,7 +3,7 @@ package com.filezipper.iostreams;
 import java.io.IOException;
 
 public interface IInputStream {
-    public Byte read() throws IOException;
+    public int read() throws IOException;
     public Byte[] readNBytes(Integer numberOfBytes) throws IOException;
     public Integer  available() throws IOException;
     public void reset() throws IOException;
