+++ src/main/java/com/filezipper/iostreams/IOutputStream.java
@@ -3,7 +3,7 @@ package com.filezipper.iostreams;
 import java.io.IOException;
 
 public interface IOutputStream {
-    public void write(Byte data) throws IOException;
+    public void write(int data) throws IOException;
     public void write(Byte[] data) throws IOException;
     public void finalize() throws Throwable;
 }
